split data into modules
