+++ modules/data.js
@@ -1,0 +1,11 @@
+export const getBooks = () => {
+  let books;
+  if (localStorage.getItem('books') === null) {
+    books = [];
+  } else {
+    books = JSON.parse(localStorage.getItem('books'));
+  }
+  return books;
+};
+
+export const saveBooks = (books) => localStorage.setItem('books', JSON.stringify(books));
